+++ src/hooks/useFetchGroupMembers.ts
@@ -1,5 +1,6 @@
 import { useState, useEffect } from 'react';
 import api from '@/apis/axiosInstance';
+import { useNavigate } from "react-router-dom";
 
 interface Emoji {
   emojiType: string;
@@ -17,6 +18,7 @@ const useFetchGroupMembers = () => {
   const [members, setMembers] = useState<Member[]>([]);
   const [loading, setLoading] = useState(true);
   const [error, setError] = useState<string | null>(null);
+  const navigate = useNavigate();
 
   useEffect(() => {
     const fetchMembers = async () => {
@@ -26,12 +28,16 @@ const useFetchGroupMembers = () => {
       try {
         const response = await api.get('/api/groups/today');
 
+
         if (response.data && response.data.isSuccess) {
           console.log("✅ API 응답 데이터:", response.data);
-          const formattedMembers = response.data.data.map((member: Member) => ({
-            ...member,
-            nickname: member.nickname || "알 수 없음",
-          }));
+          const formattedMembers = response.data.data.map((member: Member) => {
+            const storedNickname = localStorage.getItem("nickname") || "알 수 없음"; 
+            return {
+              ...member,
+              nickname: member.nickname || storedNickname, // 닉네임이 없으면 로컬 스토리지 값 사용
+            };
+          });
           setMembers(formattedMembers);
         } else {
           setError(response.data?.message || '데이터를 가져오는 데 실패했습니다.');
@@ -51,7 +57,7 @@ const useFetchGroupMembers = () => {
           if (errorResponse.response.status === 401 || errorResponse.response.status === 403) {
             console.warn("🔑 인증 오류 발생 - 로그인 필요");
             localStorage.removeItem('access_token');
-            window.location.href = '/login'; // 로그인 페이지로 리디렉트
+            navigate('/login');
           }
         } else {
           setError('서버와 연결할 수 없습니다. 네트워크 상태를 확인하세요.');
